Update work_lights.py
Catch ConnectionResetError and retry after sleeping for a minute.
Turn on lights only if script is running during work day.
Only get 3 events (the next earliest one) from Calendar. This is in case there are 3 events, all starting at the same time but there is only one that has reminders and I've accepted.